Added a User Friendly multi select interface
The multi select interface allows you easily initialize the new interface from a existing old multi select input.

Minor editing to existing select menus is required.

a script with src to `/multiselect.js` is required to prepare the new ui.

<select> tags must now have a class "multiselect" for the new ui to take place.

at the end of the page, you must add a `window.initializeMulti()` function to initialize the ui.

If anything is incorrect, please comment.