Refactor guildMemberRemove event to use global.botModel and fix bot deletion (#56)
* feat: Add global.botModel to guildMemberRemove event

The global botModel is now imported into guildMemberRemove event
for easy access to the bot model across different files in the codebase.

* docs: Add guildMemberAdd and guildMemberRemove events

The welcome message for `guildMemberAdd` event has been updated to provide new members with the necessary information regarding the server channels, rules, and support. Additionally, the team has refactored the code for the `guildMemberRemove` event and fixed a bug that prevented it from finding the right bots to remove from the database. Finally, updated the `config.js` file by including new `rules`, `info`, and `support` channel ids.

* Refactor guildMemberRemove event to use global.botModel and fix bot deletion

- Replaced Bot.findOne with global.botModel.findOne
- Changed Bot.deleteOne({ id: bot.id }) to Bot.deleteOne({ owner: member.id }) since deleting the bot owner is the correct operation.
- Moved botMember.kick() below Bot.deleteOne to ensure correct bot deletion.