Add error logging to webhooks
This commit adds a new `webhooks` object to `config.js` with a sub-object for `error_logs`. A new event listener is added to index.js that sends an embed to the specified webhook upon receiving an uncaught exception, an unhandled rejection, or an uncaught exception monitor with data and formatting included.