Some required changes (#60)
* Add my ID to config as for esting purposes I might require use of eval command

* Add error logging to webhooks

This commit adds a new `webhooks` object to `config.js` with a sub-object for `error_logs`. A new event listener is added to index.js that sends an embed to the specified webhook upon receiving an uncaught exception, an unhandled rejection, or an uncaught exception monitor with data and formatting included.

* Add placeholder comment to caching.js

* feat: Add emoji formatting to bot-related embeds

This commit modifies certain embed contents to use formatted emoji rather than explicit texts. Specifically, this change adds support for `formatEmoji` in the bot added/edited and certification requested embed descriptions.