Minor fix (#62)
* Add my ID to config as for esting purposes I might require use of eval command

* Add error logging to webhooks

This commit adds a new `webhooks` object to `config.js` with a sub-object for `error_logs`. A new event listener is added to index.js that sends an embed to the specified webhook upon receiving an uncaught exception, an unhandled rejection, or an uncaught exception monitor with data and formatting included.

* Add placeholder comment to caching.js

* feat: Add emoji formatting to bot-related embeds

This commit modifies certain embed contents to use formatted emoji rather than explicit texts. Specifically, this change adds support for `formatEmoji` in the bot added/edited and certification requested embed descriptions.

* refactor: Simplify webhook client initialization

This commit simplifies webhook client initialization by passing an object `{url: url}` to the constructor instead of just the URL in string format.

* Add error log webhooks to the configuration file

This commit adds a new webhook configuration option to the `config.js` file for handling error logs. This option is now enabled by default and will send webhook notifications to the URL specified in the `errorLogs` environment variable.