Refactor guildMemberRemove event to avoid error when kicking bots (#63)
* Add my ID to config as for esting purposes I might require use of eval command

* Add error logging to webhooks

This commit adds a new `webhooks` object to `config.js` with a sub-object for `error_logs`. A new event listener is added to index.js that sends an embed to the specified webhook upon receiving an uncaught exception, an unhandled rejection, or an uncaught exception monitor with data and formatting included.

* Add placeholder comment to caching.js

* feat: Add emoji formatting to bot-related embeds

This commit modifies certain embed contents to use formatted emoji rather than explicit texts. Specifically, this change adds support for `formatEmoji` in the bot added/edited and certification requested embed descriptions.

* refactor: Simplify webhook client initialization

This commit simplifies webhook client initialization by passing an object `{url: url}` to the constructor instead of just the URL in string format.

* Add error log webhooks to the configuration file

This commit adds a new webhook configuration option to the `config.js` file for handling error logs. This option is now enabled by default and will send webhook notifications to the URL specified in the `errorLogs` environment variable.

* Toggle `repliedUser` to true for better replied user mention parsing

In this commit, we enabled better parsing for replied user mentions by switching `repliedUser` to true. This will allow for more accurate mentions when a user is replied to, as well as better parsing for the `allowedMentions` option in the `client` object.

* Refactor guildMemberRemove event to avoid error when kicking bots

This commit updates the guildMemberRemove event code to fix an issue where an error would occur when trying to access the bots array length. Using the find() method instead of findOne() fixes this issue, ensuring a smooth bot removal process.