update developer role ID (#70)
* update developer role ID

By mistake you made the poeple with role bottester to use yuor bot's eval command please merge this pull request asap

* Refactor developer IDs in config.js

This commit changes the IDs of two developers in the owners and developers arrays. One developer is added to the developers array, and a developer ID is updated in the role IDs object.