Refactor code for kicking bots when their owner leaves the server. Bot deletion now uses global model. (#72)
* update developer role ID

By mistake you made the poeple with role bottester to use yuor bot's eval command please merge this pull request asap

* Refactor developer IDs in config.js

This commit changes the IDs of two developers in the owners and developers arrays. One developer is added to the developers array, and a developer ID is updated in the role IDs object.

* Refactor guildMemberRemove.js to use Discord.js v13 GuildMemberManager API

The code changes impact the guildMemberRemove.js file and use the GuildMemberManager class of Discord.js v13. This commit message summarizes the changes made to the code file.

* Refactor code for kicking bots when their owner leaves the server. Bot deletion now uses global model.