Add visa2discord dependency and revise eval command (#74)
* Refactor: Get member tag as a string in guildMemberRemove.js

* Add visa2discord dependency and revise eval command

The commit adds a new dependency named "visa2discord" and updates the `eval` command to make use of the new dependency, improving code security. Instead of using `eval-helper`, it now uses node's native `eval` function to get evaluated output, avoiding specific string manipulations that could make the bot vulnerable. Additionally, the commit renames `evaled` to `code` to improve readability, and creates a new `splitDescription` variable to split the output string into multiple messages if needed.

* refactor: Update Visa2Discord dependency version

This change updates the version of the Visa2Discord dependency to the latest release. Also, it adds a new function `cleanCode()` to sanitize the evaluation output before logging to the console to prevent token leakage.